requirements.txt
fastapi
uvicorn[standard]
pydantic
litellm
anthropic
openai
redis
httpx
docker
python-dotenv
rich
typer
